+++ packages/squire/package.json
@@ -4,5 +4,9 @@
 	"dependencies": {
 		"duckdb-async": "^1.1.1",
 		"toolbox": "workspace:*"
+	},
+	"scripts": {
+		"test": "vitest --run",
+		"test:watch": "vitest"
 	}
 }
